fix spread function

+++ src/timerwheel.rs
@@ -335,7 +335,7 @@ mod tests {
         let mut rng = rand::thread_rng();
         for _ in 0..50000 {
             let expire = now + Duration::from_secs(rng.gen_range(5..250)).as_nanos() as u64;
-            core.set(vec![(rng.gen_range(0..10000), expire)]);
+            core.set(vec![(rng.gen_range(0..10000), expire as i64)]);
         }
 
         for dt in [5, 6, 7, 10, 15, 20, 25, 50, 51, 52, 53, 70, 75, 85, 100] {
@@ -348,7 +348,7 @@ mod tests {
         let now = core.wheel.clock.now_ns();
         for _ in 0..10000 {
             let expire = now + Duration::from_secs(rng.gen_range(110..250)).as_nanos() as u64;
-            core.set(vec![(rng.gen_range(0..1000), expire)]);
+            core.set(vec![(rng.gen_range(0..1000), expire as i64)]);
         }
         for dt in [5, 6, 7, 10, 15, 20, 25, 50, 51, 52, 53, 70, 75, 85, 100] {
             core.wheel.advance(
